+++ tests/framework/vault/pkg_invalid/invalid.py
@@ -1,1 +1,4 @@
+#!/usr/bin/python
+# -- Content-Encoding: UTF-8 --
+
 raise ImportError("Invalid bundle")